fix: accept fee-haircut buy fills and record actual fill size
Market BUYs settle with fewer tokens than requested because Predict.fun
deducts the taker fee in the output token and MARKET fills can sweep
worse-priced asks within the slippage budget. A request for 10 shares
typically delivers ~9.83 on-chain, so the strict ">= shares" balance
check rejected real fills, leaving orphaned positions with no GTC sell.

Confirm the buy when the on-chain balance grows by at least
PREDICTFUN_BUY_CONFIRM_MIN_FILL_RATIO * shares (default 0.97, env-tunable),
then record the observed balance delta as the position's actual share
count so the GTC sell, P&L, and Slack alert all reflect what was
actually bought.

+++ trade_db.py
@@ -136,7 +136,7 @@ def insert_position(
     tick_size: float,
     buy_price: float,
     buy_fee: float,
-    shares: int,
+    shares: float,
     target_sell_price: float,
     model_prob: float,
     buy_order_id: str,
